+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/Settings/FrameSettings.cs
@@ -246,7 +246,7 @@ namespace UnityEngine.Rendering.HighDefinition
     partial struct FrameSettings
     {
         /// <summary>Default FrameSettings for Camera renderer.</summary>
-        [Obsolete("Since 2020.1, use FrameSettings.NewDefaultCamera() instead.")]
+        [Obsolete("Since 2019.3, use FrameSettings.NewDefaultCamera() instead.")]
         internal static readonly FrameSettings defaultCamera = default;
         /// <summary>Default FrameSettings for Camera renderer.</summary>
         internal static FrameSettings NewDefaultCamera() => new FrameSettings()
@@ -315,7 +315,7 @@ namespace UnityEngine.Rendering.HighDefinition
             lodBias = 1,
         };
         /// <summary>Default FrameSettings for realtime ReflectionProbe/PlanarReflectionProbe renderer.</summary>
-        [Obsolete("Since 2020.1, use FrameSettings.NewDefaultRealtimeReflectionProbe() instead.")]
+        [Obsolete("Since 2019.3, use FrameSettings.NewDefaultRealtimeReflectionProbe() instead.")]
         internal static readonly FrameSettings defaultRealtimeReflectionProbe = default;
         /// <summary>Default FrameSettings for realtime ReflectionProbe/PlanarReflectionProbe renderer.</summary>
         internal static FrameSettings NewDefaultRealtimeReflectionProbe() => new FrameSettings()
@@ -368,7 +368,7 @@ namespace UnityEngine.Rendering.HighDefinition
             lodBias = 1,
         };
         /// <summary>Default FrameSettings for baked or custom ReflectionProbe renderer.</summary>
-        [Obsolete("Since 2020.1, use FrameSettings.NewDefaultCustomOrBakeReflectionProbe() instead.")]
+        [Obsolete("Since 2019.3, use FrameSettings.NewDefaultCustomOrBakeReflectionProbe() instead.")]
         internal static readonly FrameSettings defaultCustomOrBakeReflectionProbe = default;
         /// <summary>Default FrameSettings for baked or custom ReflectionProbe renderer.</summary>
         internal static FrameSettings NewDefaultCustomOrBakeReflectionProbe() => new FrameSettings()