+++ com.unity.render-pipelines.high-definition/Runtime/Utilities/VolumeComponentWithQuality.cs
@@ -5,7 +5,7 @@ namespace UnityEngine.Rendering.HighDefinition
     public abstract class VolumeComponentWithQuality : VolumeComponent
     {
         [Tooltip("Specifies the quality level to be used for performance relevant parameters.")]
-        public ScalableSettingLevelParameter quality = new ScalableSettingLevelParameter((int)ScalableSettingLevelParameter.Level.Medium, true);
+        public ScalableSettingLevelParameter quality = new ScalableSettingLevelParameter((int)ScalableSettingLevelParameter.Level.Medium, false);
 
         static protected GlobalPostProcessingQualitySettings GetPostProcessingQualitySettings()
         {
@@ -13,6 +13,17 @@ namespace UnityEngine.Rendering.HighDefinition
             if(pipeline != null)
             {
                 return pipeline.currentPlatformRenderPipelineSettings.postProcessQualitySettings;
+            }
+            // This shouldn't happen ever.
+            return null;
+        }
+
+        static protected GlobalLightingQualitySettings GetLightingQualitySettings()
+        {
+            var pipeline = (HDRenderPipeline)RenderPipelineManager.currentPipeline;
+            if (pipeline != null)
+            {
+                return pipeline.currentPlatformRenderPipelineSettings.lightingQualitySettings;
             }
             // This shouldn't happen ever.
             return null;