+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -208,6 +208,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed Prefab override workflow issue in DensityVolume.
 - Fixed empty MoreOptions and moved AdvancedManipulation in a dedicated location for coherency in DensityVolume.
 - Fix light limit counts specified on the HDRP asset
+- Fixed Quality Settings for SSR, Contact Shadows and Ambient Occlusion volume components
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled